requirements.txt
# Django and API dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-decouple==3.8
django-filter==23.3

# Authentication
djangorestframework-simplejwt==5.3.0

# AI/ML dependencies
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
celery==5.3.4
redis==5.0.1

# Task automation
APScheduler==3.10.4
requests==2.31.0

# Development tools
python-dotenv==1.0.0

# PDF generation and reporting
reportlab==4.0.7

# Additional packages for production
gunicorn==21.2.0
whitenoise==6.6.0
setuptools>=65.0.0
